batctl: Add settings_data hooks for netlink integration
The generic netlink infrastructure will be used in the future to replace
sysfs for manipulating the runtime configuation of batman-adv meshifs,
hardifs and vlans. These will not use the raw strings when communicating
with the kernel interface but a well defined binary message format.

This means that the read function for settings must parse the binary format
and convert it to a textual representation for the user. The netlink_get
hook of struct settings_data will be used for that.

A similar problem is the setting of configuration entries. The textual
representation of the user input has to be parsed and validated. And the
resulting attributes have to be put in a message which the kernel can
interpret. The parsing is done in the parse hook which can use the data
pointer to store the results. The netlink_set hook has to prepare the
generic netlink message and send it to the kernel.

Signed-off-by: Sven Eckelmann <sven@narfation.org>