batctl: Add helper to read/write boolean configuration values
Most of the available settings commands in batctl only operate on a single
boolean value. Both the reading, parsing and writing of these values can
mostly be handled by the same set of helper functions. Only the actual
setting/getting of the correct attribute has to be handled by the actual
settings command implementation.

Signed-off-by: Sven Eckelmann <sven@narfation.org>